+++ test/test.refresh.ts
@@ -1,21 +1,20 @@
-/**
- * Copyright 2013 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2013 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 import * as assert from 'assert';
 import * as fs from 'fs';
+import * as nock from 'nock';
 import {UserRefreshClient} from '../src';
 
 // Creates a standard JSON credentials object for testing.
@@ -122,4 +121,22 @@ it('fromStream should read the stream and create a UserRefreshClient', done => {
     assert.strictEqual(json.refresh_token, refresh._refreshToken);
     done();
   });
+});
+
+it('getRequestHeaders should populate x-goog-user-project header if quota_project_id present', async () => {
+  // The first time auth.getRequestHeaders() is called /token endpoint is used to
+  // fetch a JWT.
+  const req = nock('https://oauth2.googleapis.com')
+    .post('/token')
+    .reply(200, {});
+
+  // Fake loading default credentials with quota project set:
+  const stream = fs.createReadStream(
+    './test/fixtures/config-with-quota/.config/gcloud/application_default_credentials.json'
+  );
+  const refresh = new UserRefreshClient();
+  await refresh.fromStream(stream);
+
+  const headers = await refresh.getRequestHeaders();
+  assert.strictEqual(headers['x-goog-user-project'], 'my-quota-project');
 });