+++ test/test.transporters.ts
@@ -1,18 +1,16 @@
-/**
- * Copyright 2013 Google Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2013 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 import * as assert from 'assert';
 import {GaxiosOptions} from 'gaxios';
@@ -71,9 +69,25 @@ it('should append to x-goog-api-client header if it exists', () => {
     headers: {'x-goog-api-client': 'gdcl/1.0.0'},
     url: '',
   });
-  console.info(opts.headers);
   assert(
     /^gdcl\/[.-\w$]+ auth\/[.-\w$]+$/.test(opts.headers!['x-goog-api-client'])
+  );
+});
+
+// see: https://github.com/googleapis/google-auth-library-nodejs/issues/819
+it('should not append x-goog-api-client header multiple times', () => {
+  const opts = {
+    headers: {'x-goog-api-client': 'gdcl/1.0.0'},
+    url: '',
+  };
+  let configuredOpts = transporter.configure(opts);
+  console.info(configuredOpts);
+  configuredOpts = transporter.configure(opts);
+  console.info(configuredOpts);
+  assert(
+    /^gdcl\/[.-\w$]+ auth\/[.-\w$]+$/.test(
+      configuredOpts.headers!['x-goog-api-client']
+    )
   );
 });